+++ sdk/keyvault/Azure.Security.KeyVault.Administration/src/KeyVaultAccessControlClient.cs
@@ -67,7 +67,7 @@ namespace Azure.Security.KeyVault.Administration
                 perRetryPolicies: [new ChallengeBasedAuthenticationPolicy(
                     credential,
                     options.DisableChallengeResourceVerification,
-                    ChallengeBasedAuthenticationPolicy.SupportsProofOfPossession(options.Transport))],
+                    options.EnableProofOfPossession && ChallengeBasedAuthenticationPolicy.SupportsProofOfPossession(options.Transport))],
                 transportOptions: new HttpPipelineTransportOptions(),
                 responseClassifier: null);