+++ sdk/keyvault/Azure.Security.KeyVault.Keys/src/Cryptography/RemoteCryptographyClient.cs
@@ -35,7 +35,7 @@ namespace Azure.Security.KeyVault.Keys.Cryptography
                 perRetryPolicies: [new ChallengeBasedAuthenticationPolicy(
                     credential,
                     options.DisableChallengeResourceVerification,
-                    ChallengeBasedAuthenticationPolicy.SupportsProofOfPossession(options.Transport))],
+                    options.EnableProofOfPossession && ChallengeBasedAuthenticationPolicy.SupportsProofOfPossession(options.Transport))],
                 transportOptions: new HttpPipelineTransportOptions(),
                 responseClassifier: null);