+++ sdk/keyvault/Azure.Security.KeyVault.Secrets/src/SecretClient.cs
@@ -68,7 +68,7 @@ namespace Azure.Security.KeyVault.Secrets
                 perRetryPolicies: [new ChallengeBasedAuthenticationPolicy(
                     credential,
                     options.DisableChallengeResourceVerification,
-                    ChallengeBasedAuthenticationPolicy.SupportsProofOfPossession(options.Transport))],
+                    options.EnableProofOfPossession && ChallengeBasedAuthenticationPolicy.SupportsProofOfPossession(options.Transport))],
                 transportOptions: new HttpPipelineTransportOptions(),
                 responseClassifier: null);
             _pipeline = pipeline;